// ui3/src/generated/bridge/TravelStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Operation } from "./Operation";
import type { TravelPhase } from "./TravelPhase";

export type TravelStatus = { operation: Operation, realmOperation: Operation | null, phase: TravelPhase, realm: string, parcel: [number, number] | null, blockingReason: string | null, canCancel: boolean, };